Create Export_to_JSON.py
This script will export the selected objects mesh in to a JSON formatted file. 
The mesh must not have only triangles. 
To Do:
Apply the triangulate modifier in order to make sure the mesh is contains only triangles,
Add the option to export copies of the mesh represented by the transforms: rotation translation and scale